Revert "Feature/psi j"
