+++ tests/unit_tests/test_launcher/test_launcher.py
@@ -11,7 +11,7 @@ from unittest import mock
 import radical.utils as ru
 import radical.pilot as rp
 
-from   radical.pilot.pmgr.launching.base import PMGRLaunchingComponent
+from   radical.pilot.pmgr.launching.default import Default
 
 
 # ------------------------------------------------------------------------------
@@ -54,12 +54,12 @@ class TestLauncher(TestCase):
 
     # --------------------------------------------------------------------------
     #
-    @mock.patch.object(PMGRLaunchingComponent, '__init__', return_value=None)
+    @mock.patch.object(Default, '__init__', return_value=None)
     def test_configure(self, mocked_init):
 
         session, configs = self.setUp()
 
-        component = PMGRLaunchingComponent(cfg=None, session=None)
+        component = Default(cfg=None, session=None)
         component._uid        = 'pmgr.launching.0000'
         component._cfg        = mock.Mock()
         component._log        = ru.Logger('dummy')
@@ -68,10 +68,13 @@ class TestLauncher(TestCase):
 
         component._pmgr       = 'pmgr.0'
         component._prof       = ru.Config(cfg = {'enabled': False})
+        component._cache_lock = ru.Lock()
+        component._cache      = dict()
         component._sandboxes  = dict()
 
         component._mod_dir    = os.path.dirname(os.path.abspath(__file__))
         component._root_dir   = "%s/../../src/radical/pilot/" % component._mod_dir
+        component._conf_dir   = "%s/configs/" % component._root_dir
 
         component._rp_version    = rp.version
         component._rp_sdist_name = rp.sdist_name
@@ -95,9 +98,9 @@ class TestLauncher(TestCase):
                                          'candidate_hosts': None,
                                          'services'       : []}
                    }
-        component._prepare_pilot(resource, rcfg, pilot, {}, '')
-        self.assertEqual(pilot['jd_dict'].name, 'pilot.0001')
-        self.assertEqual(pilot['jd_dict'].environment['RADICAL_BASE'],
+        ret = component._prepare_pilot(resource, rcfg, pilot, {}, '')
+        self.assertEqual(ret['jd'].name, 'pilot.0001')
+        self.assertEqual(ret['jd'].environment['RADICAL_BASE'],
                          str(session._get_resource_sandbox(pilot)))
 
         pilot    = {
@@ -115,9 +118,9 @@ class TestLauncher(TestCase):
                                          'candidate_hosts': None,
                                          'services'       : []}
                    }
-        component._prepare_pilot(resource, rcfg, pilot, {}, '')
-        self.assertEqual(pilot['jd_dict'].name, 'pilot.0000')
-        self.assertEqual(pilot['jd_dict'].environment['RADICAL_BASE'],
+        ret = component._prepare_pilot(resource, rcfg, pilot, {}, '')
+        self.assertEqual(ret['jd'].name, 'pilot.0000')
+        self.assertEqual(ret['jd'].environment['RADICAL_BASE'],
                          str(session._get_resource_sandbox(pilot)))
 
         pilot    = {
@@ -135,17 +138,19 @@ class TestLauncher(TestCase):
                                          'candidate_hosts': None,
                                          'services'       : []}
                    }
-        component._prepare_pilot(resource, rcfg, pilot, {}, '')
-        self.assertEqual(pilot['jd_dict'].name, 'bar')
+        ret = component._prepare_pilot(resource, rcfg, pilot, {}, '')
+        self.assertEqual(ret['jd'].name, 'bar')
+
+        # test resource config parameter `system_architecture`
 
         # default value is {}
-        self.assertEqual(pilot['jd_dict'].system_architecture, {})
+        self.assertEqual(ret['jd'].system_architecture, {})
 
         # value for "ornl.summit" is 4
         resource = 'ornl.summit'
         rcfg = configs.ornl.summit
-        component._prepare_pilot(resource, rcfg, pilot, {}, '')
-        self.assertEqual(pilot['jd_dict'].system_architecture['smt'], 4)
+        ret = component._prepare_pilot(resource, rcfg, pilot, {}, '')
+        self.assertEqual(ret['jd'].system_architecture['smt'], 4)
 
 
 # ------------------------------------------------------------------------------
